requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
configparser==5.3.0
findspark==2.0.1
idna==3.4
kafka-python==1.4.7
oauthlib==3.2.2
psutil==5.9.4
py4j==0.10.7
PySocks==1.7.1
pyspark==2.4.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.12.1
urllib3==1.26.14